voice-service/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
faster-whisper>=1.0.0
elevenlabs>=1.0.0
openai>=1.0.0
piper-tts>=1.2.0
pydub>=0.25.0
python-dotenv>=1.0.0
httpx>=0.25.0
numpy>=1.24.0
